refactor(gui): move the Codex strategy write beside the client it uses
account-pool-strategy.ts owns the value normalizers pool-settings.ts imports, so keeping the transport there too made the two modules import each other. The first draft hid that behind a dynamic import and the bundler reported INEFFECTIVE_DYNAMIC_IMPORT - the cycle telling on itself. account-pool-strategy.ts is a pure value module again.
